models/requirements.txt
openai
openai-whisper @ git+https://github.com/Blair-Johnson/batch-whisper.git@main
numpy
redis
pydub
python-dotenv